requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.6
channels==3.0.5
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==5.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.0
zope.interface==5.4.0
redis==3.5.3
pillow==8.2.0
pymysql==1.0.2
